requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0
django-cors-headers==2.2.0
djangorestframework==3.9.1
gunicorn==19.8.1
idna==2.8
kombu==4.3.0
Pillow==5.1.0
psycopg2==2.7.7
pycryptodome==3.8.0
PyJWT==1.6.1
PySocks==1.6.8
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.9
rave-python==1.0.1
cloudinary==1.15.0
redis==3.1.0
requests==2.21.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.12.0
twilio==6.24.0
urllib3==1.24.1
vine==1.2.0
whitenoise==3.3.1
